requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
pluggy==1.2.0
pycparser==2.21
PySocks==1.7.1
pytest==7.4.0
PyYAML==6.0.1
requests==2.31.0
selenium==4.11.2
selenium-stealth==1.0.6
sniffio==1.3.0
sortedcontainers==2.4.0
stealth==0.0.1
trio==0.22.2
trio-websocket==0.10.3
undetected-chromedriver==3.5.3
urllib3==1.26.16
websockets==11.0.3
wsproto==1.2.0
yml==0.0.1
